+++ apps/exams/views/admin/deployments_detail.py
@@ -10,7 +10,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from apps.exams.constants import ErrorMessages
-from apps.exams.permissions import IsExamStaff
+from apps.core.utils.permissions import IsStaffRole
 from apps.exams.serializers.admin.deployments_detail import (
     AdminExamDeploymentDetailResponseSerializer,
 )
@@ -24,7 +24,7 @@ from apps.exams.views.mixins import ExamsExceptionMixin
 class AdminExamDeploymentDetailAPIView(ExamsExceptionMixin, APIView):
     """어드민 쪽지시험 배포 상세 조회 API."""
 
-    permission_classes = [IsAuthenticated, IsExamStaff]
+    permission_classes = [IsAuthenticated, IsStaffRole]
     serializer_class = AdminExamDeploymentDetailResponseSerializer
 
     def permission_denied(self, request: Request, message: str | None = None, code: str | None = None) -> NoReturn:
