+++ apps/exams/views/admin/deployments_status.py
@@ -9,7 +9,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from apps.exams.constants import ErrorMessages
-from apps.exams.permissions import IsExamStaff
+from apps.core.utils.permissions import IsStaffRole
 from apps.exams.serializers.admin.deployments_status import (
     AdminExamDeploymentStatusRequestSerializer,
     AdminExamDeploymentStatusResponseSerializer,
@@ -25,7 +25,7 @@ from apps.exams.views.mixins import ExamsExceptionMixin
 class AdminExamDeploymentStatusAPIView(ExamsExceptionMixin, APIView):
     """어드민 쪽지시험 배포 상태 변경 API."""
 
-    permission_classes = [IsAuthenticated, IsExamStaff]
+    permission_classes = [IsAuthenticated, IsStaffRole]
     serializer_class = AdminExamDeploymentStatusRequestSerializer
 
     def permission_denied(self, request: Request, message: str | None = None, code: str | None = None) -> NoReturn:
