+++ apps/exams/views/admin/exams_delete.py
@@ -8,7 +8,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from apps.exams.constants import ErrorMessages
-from apps.exams.permissions import IsExamStaff
+from apps.core.utils.permissions import IsStaffRole
 from apps.exams.serializers.admin.exams_delete import (
     AdminExamDeleteResponseSerializer,
 )
@@ -23,7 +23,7 @@ from apps.exams.views.mixins import ExamsExceptionMixin
 class AdminExamDeleteAPIView(ExamsExceptionMixin, APIView):
     """어드민 쪽지시험 삭제 API."""
 
-    permission_classes = [IsAuthenticated, IsExamStaff]
+    permission_classes = [IsAuthenticated, IsStaffRole]
     serializer_class = AdminExamDeleteResponseSerializer
 
     def permission_denied(self, request: Request, message: str | None = None, code: str | None = None) -> NoReturn:
