+++ apps/exams/views/student/deployments_status.py
@@ -6,7 +6,7 @@ from rest_framework.views import APIView
 
 from apps.exams.constants import ErrorMessages, ExamStatus
 from apps.exams.models import ExamDeployment
-from apps.exams.permissions import IsStudentRole
+from apps.core.utils.permissions import IsStudentRole
 from apps.exams.serializers.student.deployments_status import (
     ExamStatusResponseSerializer,
 )
